requirements.txt
pyyaml~=6.0
pydantic~=1.10